+++ MICA/Código_Sensores/AS5600_Magnetic_Encoder/platformio.ini
@@ -1,0 +1,6 @@
+[env:esp8266]
+platform = espressif8266
+board = nodemcuv2     ; o d1_mini si usas un Wemos D1 mini
+framework = arduino
+monitor_speed = 115200
+upload_speed = 115200
