+++ MICA/Código_Sensores/GY_273_QMC5883L_Compass/platformio.ini
@@ -1,0 +1,14 @@
+[env:esp8266]
+platform = espressif8266
+board = nodemcuv2
+framework = arduino
+monitor_speed = 115200
+lib_deps = 
+  mprograms/QMC5883LCompass @ ^1.2.3
+
+
+[env:esp32-c6-devkitm-1]
+platform  = https://github.com/pioarduino/platform-espressif32/releases/download/51.03.04/platform-espressif32.zip
+board = esp32-c6-devkitm-1
+framework = arduino
+monitor_speed = 115200
