+++ MICA/Data_Analysis/README.md
@@ -1,0 +1,27 @@
+# Dades MICA
+
+Aquesta carpeta proporciona dades i scripts per ajudar els participants a crear solucions innovadores per promoure l'ús responsable de l'aigua.
+
+## Contingut del repositori
+
+- **Dades**: Conjunts de dades reals del consum d'aigua d'una llar durant una setmana:
+    - `water_consumption_by_second.csv`: Conté el consum d'aigua total de la llar per cada segon.
+    - `water_consumption_by_minute.csv`: Conté el consum d'aigua total de la llar per cada minut.
+    - `raw_sensor_data.csv`: Conté les dades que rebem del sensor directament, on es genera una fila cada vegada que el caudal d'aigua canvia. 
+    - `event_labels.csv`: Conté classificacions manuals fetes pels habitants de la llar, en què han indicat l'inici, el final i la categoria dels events de consum que han realitzat.
+- **Notebooks**: Notebooks en Python per ajudar els participants a analitzar i visualitzar les dades, en formats `ipynb`, `pdf`, `html` i els gràfics generats.  
+
+## Com començar
+
+1. Clona aquest repositori  
+2. Explora la carpeta `data/` per als conjunts de dades i la carpeta `notebooks/` per a les eines d'anàlisi.
+
+## Contribuir
+
+Les contribucions són benvingudes! No dubtis a enviar pull requests o obrir issues per millorar aquest repositori.
+
+## Llicència
+
+Aquest projecte està llicenciat sota la [Llicència MIT](LICENSE).
+
+Bona sort, i fem un canvi plegats! 🌊
