+++ MICA/Data_Analysis/data/event_labels.csv
@@ -1,0 +1,46 @@
+"category","tag","start_timestamp","end_timestamp"
+Grifo,"",2025-03-26 19:23:29.825 +0100,2025-03-26 19:23:52.072 +0100
+Grifo,"",2025-03-26 19:24:11.621 +0100,2025-03-26 19:24:49.803 +0100
+Grifo,"",2025-03-26 19:25:20.283 +0100,2025-03-26 19:25:39.161 +0100
+Grifo,"",2025-03-26 19:26:19.152 +0100,2025-03-26 19:28:06.471 +0100
+Grifo,"",2025-03-26 19:49:45.658 +0100,2025-03-26 19:50:20.381 +0100
+Grifo,"",2025-03-26 19:50:53.658 +0100,2025-03-26 19:51:09.181 +0100
+Grifo,"",2025-03-26 19:51:23.135 +0100,2025-03-26 19:51:35.214 +0100
+Cisterna,#Mamá,2025-03-26 19:52:02.819 +0100,2025-03-26 19:52:26.086 +0100
+Cisterna,#Hijo,2025-03-26 19:53:22.564 +0100,2025-03-26 19:53:55.714 +0100
+Cisterna,#Hijo,2025-03-26 19:54:00.104 +0100,2025-03-26 19:55:10.170 +0100
+Cisterna,#Papá,2025-03-26 19:55:38.464 +0100,2025-03-26 19:56:38.858 +0100
+Cisterna,#Papá,2025-03-26 19:56:42.738 +0100,2025-03-26 19:57:41.952 +0100
+Grifo,"",2025-03-26 22:27:16.237 +0100,2025-03-26 22:27:26.538 +0100
+Ducha,#Mamá,2025-03-27 06:06:47.967 +0100,2025-03-27 06:11:19.173 +0100
+Ducha,#Papá,2025-03-27 07:30:32.701 +0100,2025-03-27 07:34:38.051 +0100
+Cisterna,#Hijo,2025-03-27 18:56:45.567 +0100,2025-03-27 18:57:12.991 +0100
+Grifo,"",2025-03-27 19:15:26.515 +0100,2025-03-27 19:15:43.638 +0100
+Lavadora,"",2025-03-27 22:25:18.400 +0100,2025-03-27 23:19:24.687 +0100
+Cisterna,#Papá,2025-03-27 22:49:30.124 +0100,2025-03-27 22:50:26.727 +0100
+Cisterna,#Hijo,2025-03-27 23:54:25.349 +0100,2025-03-27 23:56:16.424 +0100
+Grifo,"",2025-03-27 23:55:13.743 +0100,2025-03-27 23:55:16.211 +0100
+Ducha,#Hijo,2025-03-27 23:57:48.768 +0100,2025-03-28 00:04:35.987 +0100
+Grifo,"",2025-03-28 00:06:34.087 +0100,2025-03-28 00:06:56.241 +0100
+Grifo,"",2025-03-28 00:11:46.847 +0100,2025-03-28 00:12:06.577 +0100
+Cisterna,"",2025-03-28 03:02:21.701 +0100,2025-03-28 03:04:23.373 +0100
+Cisterna,#Papá,2025-03-28 06:01:59.649 +0100,2025-03-28 06:03:30.249 +0100
+Ducha,#Mamá,2025-03-28 06:07:29.603 +0100,2025-03-28 06:11:45.796 +0100
+Ducha,#Papá,2025-03-28 07:17:57.855 +0100,2025-03-28 07:21:21.430 +0100
+Cisterna,#Papá,2025-03-28 06:49:47.776 +0100,2025-03-28 06:50:46.829 +0100
+Lavaplatos,"",2025-03-28 07:34:28.333 +0100,2025-03-28 08:14:30.014 +0100
+Lavaplatos,"",2025-03-28 11:36:14.134 +0100,2025-03-28 12:24:04.908 +0100
+Cisterna,#Papá,2025-03-28 18:09:14.701 +0100,2025-03-28 18:09:36.386 +0100
+Cisterna,#Mamá,2025-03-28 18:43:52.148 +0100,2025-03-28 18:44:13.374 +0100
+Cisterna,#Papá,2025-03-29 06:19:52.026 +0100,2025-03-29 06:21:30.286 +0100
+Cisterna,"",2025-03-29 07:29:11.360 +0100,2025-03-29 07:30:33.290 +0100
+Cisterna,#Hijo,2025-03-29 07:52:31.849 +0100,2025-03-29 07:53:28.782 +0100
+Cisterna,#Papá,2025-03-31 06:27:24.095 +0200,2025-03-31 06:28:45.220 +0200
+Ducha,#Mamá,2025-03-31 06:32:48.996 +0200,2025-03-31 06:37:47.447 +0200
+Cisterna,#Papá,2025-04-01 05:44:04.293 +0200,2025-04-01 05:45:54.418 +0200
+Grifo,"",2025-04-01 21:36:16.158 +0200,2025-04-01 21:36:35.469 +0200
+Cisterna,#Papá,2025-04-02 06:06:28.806 +0200,2025-04-02 06:07:30.149 +0200
+Ducha,#Mamá,2025-04-02 06:10:24.277 +0200,2025-04-02 06:14:54.128 +0200
+Cisterna,#Papá,2025-04-03 07:12:56.290 +0200,2025-04-03 07:13:55.193 +0200
+Ducha,#Hijo,2025-04-03 11:28:27.227 +0200,2025-04-03 11:33:43.320 +0200
+Lavaplatos,"",2025-04-03 16:03:17.463 +0200,2025-04-03 16:49:35.973 +0200
