+++ MICA/Data_Analysis/requirements.txt
@@ -1,0 +1,3 @@
+plotly
+pandas
+nbformat>=4.2.0
