+++ MICA/Documentos_Soporte/README.md
@@ -1,0 +1,25 @@
+# Documentos de Soporte
+
+Esta carpeta contiene documentos de soporte relacionados con el proyecto Hackathon25Aigua. Aquí se encuentran diversos documentos que proporcionan información técnica, ideas de proyectos, estudios y análisis que apoyan el desarrollo del proyecto.
+
+## Tipos de Documentos
+
+1. **Guías Técnicas**
+   - Documentos que explican cómo implementar soluciones técnicas específicas, incluyendo el uso de sensores, optimización de energía, y más.
+
+2. **Estudios y Análisis**
+   - Informes y estudios sobre el consumo de agua, métodos de medición, y análisis de datos.
+
+3. **Ideas de Proyectos**
+   - Propuestas de proyectos que pueden ser desarrollados como parte del hackathon, incluyendo aplicaciones IoT y optimización de recursos.
+
+4. **Documentos de Trabajo**
+   - Archivos que contienen ideas y propuestas discutidas durante el hackathon, así como planes de trabajo y metodologías.
+
+## Contribuciones
+
+Si deseas contribuir a este proyecto, por favor, sigue las pautas establecidas en el archivo `CONTRIBUTING.md` del repositorio principal.
+
+## Contacto
+
+Para cualquier consulta o sugerencia, por favor contacta con los mantenedores del proyecto.
