--- MICA/Hackaton UPF Mataro/buzon de ideas/readme
@@ -1,74 +1,0 @@
-# Buzón de Ideas
-
-Bienvenido al buzón de ideas del proyecto Hackathon 25 Aigua. Este espacio está destinado a la recopilación y discusión de ideas innovadoras que ayuden a mejorar nuestro proyecto. Aquí puedes compartir tus sugerencias, propuestas y cualquier pensamiento que creas pueda contribuir al éxito del hackathon.
-
-## ¿Cómo contribuir?
-
-1. **Proponer una idea**:
-    - Crea un nuevo documento en esta carpeta con el título de tu idea como nombre del archivo.
-    - Asegúrate de incluir los siguientes detalles en el documento:
-      - **Título de la Idea**
-      - **Descripción**: Explica detalladamente la idea, incluyendo el contexto y el problema que estás tratando de resolver.
-      - **Propuesta de Solución**: Describe cómo planeas abordar el problema, incluyendo posibles estrategias, tecnologías o metodologías a utilizar.
-      - **Beneficios**: Enumera los beneficios que traería la implementación de esta idea.
-      - **Recursos Necesarios**: Indica los recursos que necesitarías para desarrollar la idea (por ejemplo, tiempo, herramientas, colaboradores).
-      - **Datos de Contacto**: Proporciona tus datos de contacto para que otros colaboradores puedan comunicarse contigo para discutir y desarrollar la idea.
-
-2. **Discutir una idea**:
-    - Revisa los documentos existentes en esta carpeta y participa en la discusión añadiendo comentarios y sugerencias directamente en el archivo.
-    - Usa reacciones en los comentarios para mostrar tu apoyo o desacuerdo con las ideas propuestas.
-
-3. **Desarrollar una idea**:
-    - Si encuentras una idea que te apasiona y quieres trabajar en ella, deja un comentario en el documento indicando tu interés y cómo planeas contribuir.
-    - Colabora con otros miembros del equipo para desarrollar la idea y llevarla a cabo.
-
-## Formato de las ideas
-
-Para mantener la organización y claridad, te sugerimos seguir el siguiente formato al proponer una nueva idea:
-
-```
-### Título de la Idea
-
-**Descripción**:
-- Explica detalladamente la idea.
-- Incluye el contexto y el problema que estás tratando de resolver.
-
-**Propuesta de Solución**:
-- Describe cómo planeas abordar el problema.
-- Incluye posibles estrategias, tecnologías o metodologías a utilizar.
-
-**Beneficios**:
-- Enumera los beneficios que traería la implementación de esta idea.
-
-**Recursos Necesarios**:
-- Indica los recursos que necesitarías para desarrollar la idea (por ejemplo, tiempo, herramientas, colaboradores).
-
-**Datos de Contacto**:
-- Proporciona tus datos de contacto para que otros colaboradores puedan comunicarse contigo para discutir y desarrollar la idea.
-
-**Comentarios Adicionales**:
-- Cualquier información adicional que consideres relevante.
-```
-
-## Normas de la Comunidad
-
-- Sé respetuoso y constructivo en tus comentarios.
-- Fomenta la colaboración y el trabajo en equipo.
-- Asegúrate de que tus propuestas sean claras y bien fundamentadas.
-- Evita duplicar ideas; revisa las propuestas existentes antes de añadir una nueva.
-
-## Contacto
-
-Si tienes alguna pregunta o necesitas más información, no dudes en contactarnos a través de los canales de comunicación del proyecto.
-
-¡Gracias por tu contribución y por ser parte de este emocionante proyecto!
-
----
-
-**Hackathon 25 Aigua - Proyecto MICA - Hackaton UPF Mataró**
-
-¡Gracias por tu contribución y por ser parte de este emocionante proyecto!
-
----
-
-**Hackathon 25 Aigua - Proyecto MICA - Hackaton UPF Mataró**
