--- MICA/Hackaton UPF Mataro/documentos soporte/Medición de Ondas de Presión para Estimar el Consumo de Agua en el Hogar.md
@@ -1,44 +1,0 @@
-# **Medición de Ondas de Presión para Estimar el Consumo de Agua en el Hogar**
-
-## **Introducción**
-
-La gestión eficiente del agua es esencial para la sostenibilidad ambiental y la reducción de costos en el hogar. La medición de ondas de presión en las tuberías domésticas ofrece una metodología innovadora para monitorear y analizar el consumo de agua, permitiendo identificar patrones de uso y detectar posibles fugas.​
-
-## **Fundamentos Teóricos**
-
-### **Dinámica de Fluidos en Tuberías**
-
-Cuando se abre un grifo en una instalación doméstica, se genera una disminución de la presión en las tuberías debido al incremento del flujo de agua hacia el punto de consumo. Esta caída de presión es resultado de la resistencia que ofrece la tubería al paso del agua; a mayor caudal, mayor pérdida de presión. Además, si se abren múltiples grifos simultáneamente, la presión en cada uno será menor, ya que el caudal total se distribuye entre más salidas. ​
-
-### **Principio de Bernoulli**
-
-El principio de Bernoulli establece que, en un flujo de fluido ideal sin fricción, la suma de la presión estática, la presión dinámica y la presión gravitacional se mantiene constante a lo largo de una línea de corriente. Aplicado a las tuberías domésticas, este principio indica que cambios en la velocidad del flujo afectan la presión dentro de la tubería. ​[Meacon Co.,Ltd](https://es.meaconsensor.com/training/flow-rate-and-pressure?utm_source=chatgpt.com)
-
-## **Aplicaciones Prácticas**
-
-### **Monitoreo del Consumo de Agua**
-
-Al instalar sensores de presión en puntos estratégicos de la red de tuberías domésticas, es posible registrar las variaciones de presión que ocurren cuando se utiliza agua en diferentes áreas o dispositivos. Analizando estas variaciones, se puede estimar el caudal y, por ende, el volumen de agua consumido en tiempo real.​
-
-### **Detección de Fugas**
-
-Las fugas en las tuberías generan patrones específicos de variación de presión. Un sistema de monitoreo continuo puede identificar estas anomalías, permitiendo la detección temprana de fugas y evitando desperdicios de agua y daños estructurales. ​
-
-### **Optimización del Uso de Agua**
-
-Con información detallada sobre el consumo de agua en diferentes áreas del hogar, los usuarios pueden tomar decisiones informadas para reducir su consumo, implementar prácticas de ahorro y optimizar el uso de recursos hídricos.​
-
-## **Consideraciones para la Implementación**
-
-* **Selección de Sensores:** Es esencial elegir sensores de presión con la precisión y sensibilidad adecuadas para detectar las variaciones de presión típicas en las tuberías domésticas.​
-
-* **Ubicación de los Sensores:** La colocación estratégica de los sensores es crucial para captar las variaciones de presión relevantes y obtener datos precisos sobre el consumo en diferentes áreas del hogar.​
-
-* **Análisis de Datos:** Desarrollar algoritmos capaces de interpretar las señales de presión y traducirlas en información útil sobre el consumo y la detección de fugas.​
-
-* **Interfaz de Usuario:** Implementar una plataforma accesible que permita a los usuarios visualizar y comprender fácilmente su consumo de agua, facilitando la toma de decisiones para el ahorro y la eficiencia.​
-
-## **Conclusión**
-
-La medición de ondas de presión en las tuberías domésticas representa una herramienta poderosa para la gestión eficiente del agua en el hogar. Al proporcionar información detallada sobre el consumo y permitir la detección temprana de fugas, esta tecnología contribuye al ahorro económico y a la sostenibilidad ambiental.
-
