--- MICA/Hackaton UPF Mataro/documentos soporte/💡 Ideas de proyectos con microprocesador en cisternas.md
@@ -1,90 +1,0 @@
-### **💡 Ideas de proyectos con microprocesador en cisternas**
-
-#### **1\. Medición del consumo de agua**
-
-* **Objetivo:** Medir cuánta agua se utiliza en cada descarga.
-
-* **Cómo:** Usar un sensor de flujo de agua conectado a un microprocesador (como un Arduino, ESP32, etc.) para registrar cuántos litros pasan durante la descarga.
-
-* **Datos que podrías obtener:**
-
-  * Promedio de litros por descarga
-
-  * Consumo diario/semanal/mensual
-
-  * Notificaciones de alto consumo
-
-#### **2\. Optimización del volumen de descarga**
-
-* **Objetivo:** Ajustar automáticamente la cantidad de agua usada según el tipo de uso (número 1 o número 2 😅).
-
-* **Cómo:**
-
-  * Sensor de presencia o botón doble (ya existe, pero se puede automatizar más)
-
-  * Uso de válvulas electrónicas controladas por el microprocesador
-
-  * Incluso se puede usar una cámara o sensor infrarrojo para detectar qué tipo de uso fue (más avanzado, pero posible)
-
-#### **3\. Control vía app o Wi-Fi**
-
-* **Microprocesadores como el ESP8266 o ESP32** permiten enviar los datos a una app o una página web local.
-
-* Puedes recibir notificaciones, ver gráficas, o activar funciones como descarga automática o modo ahorro cuando estás fuera de casa.
-
-#### **4\. Recolección de datos para análisis**
-
-* Si se instala en varios inodoros (por ejemplo, en un edificio o escuela), podrías usarlo para analizar patrones de consumo y proponer mejoras globales.
-
----
-
-### **🔧 Sensores y componentes útiles**
-
-| Componente | Función |
-| ----- | ----- |
-| Sensor de flujo YF-S201 | Mide la cantidad de agua que fluye por la tubería |
-| Solenoide de agua | Controla electrónicamente el flujo |
-| Ultrasonido HC-SR04 | Para medir el nivel del agua dentro de la cisterna |
-| ESP32 o Arduino Nano | Microprocesadores pequeños y potentes |
-| Módulo Wi-Fi | Para conectividad remota (integrado en el ESP32) |
-
----
-
-### **🛠 Ejemplos de proyectos similares (DIY)**
-
-1. **Hackaday / Instructables**
-
-   * Puedes buscar en estas plataformas con términos como:
-
-     * "Smart toilet cistern"
-
-     * "Water-saving flush system Arduino"
-
-     * "ESP32 toilet water monitor"
-
-2. **Proyectos en GitHub**
-
-   * Hay algunos con código abierto que miden el consumo de agua en general y pueden adaptarse para cisternas.
-
----
-
-### **🛒 Productos comerciales relacionados**
-
-* **Toilets inteligentes (smart toilets)**: Algunos modelos de marcas como Toto o Kohler integran sensores y modos de ahorro, aunque no todos permiten modificaciones ni muestran los datos de consumo detalladamente.
-
-* **Dispositivos retrofit**: Hay algunos gadgets que se instalan en cisternas tradicionales para convertirlas en dual flush o activar por sensor, pero son limitados en cuanto a monitoreo.
-
----
-
-### **🧠 Ideas para ir más allá**
-
-* **Integrar energía solar o usar una dínamo para autoalimentarse.**
-
-* **Recolección de agua de lluvia para usar en descargas.**
-
-* **Gamificación del consumo de agua (app con medallas por ahorrar, rankings, etc.)**
-
----
-
-Si quieres te puedo ayudar a diseñar uno con esquemas, código y materiales. ¿Tienes alguna idea más específica o quieres hacerlo para tu casa, para un proyecto, o más como producto comercial?
-
