--- MICA/Hackaton UPF Mataro/documentos soporte/📘 Documento de trabajo hackaton MICA.md
@@ -1,208 +1,0 @@
-# **📘 Documento de trabajo hackaton MICA**
-
----
-
-## **🌊 Introducción**
-
-El proyecto **MICA (Monitorització Intel·ligent del Consum d’Aigua)** propone utilizar sensores IoT e inteligencia artificial para monitorizar y optimizar el consumo de agua en hogares urbanos. Este documento está dividido en dos secciones principales:
-
-1. **Áreas de Desarrollo e Ideas Generales:**  
-    Se describen los ámbitos y ejemplos donde los equipos pueden innovar para ampliar y mejorar la solución MICA.
-
-2. **Infraestructura Básica Requerida:**  
-    Se detallan los componentes y datos esenciales que deben estar listos antes de comenzar la hackatón, priorizando aquellos que no requieren instalación por parte de los participantes.
-
----
-
-## **1\. 🚀 Áreas de Desarrollo e Ideas Generales**
-
-### **1.1 Optimización de la Captura y Transmisión de Datos**
-
-* **Descripción:**  
-   Mejorar la precisión y eficiencia en la recolección de datos del sensor de caudal.
-
-* **Ejemplo:**  
-   Desarrollo de algoritmos que filtren ruidos o interferencias y optimicen el consumo energético en el ESP32 al transmitir datos vía MQTT.
-
-* **Reto técnico sugerido:**  
-   *Implementación de sensores con imanes anulares multipolo y un Time-to-Digital Converter (TDC)*  
-   *(Desafío de medición de velocidad instantánea con AS5200L).*
-
-### **1.2 Detección de Fugas y Anomalías**
-
-* **Descripción:**  
-   Aplicar IA para detectar patrones de consumo anómalos que indiquen fugas o fallos.
-
-* **Ejemplo:**  
-   Sistema de alertas por consumo prolongado en horarios de baja actividad.
-
-### **1.3 Análisis de Patrones de Comportamiento**
-
-* **Descripción:**  
-   Comparar hábitos de consumo para identificar oportunidades de ahorro, especialmente en colectivos vulnerables.
-
-* **Ejemplo:**  
-   Uso de *clustering* para segmentar usuarios y ofrecer recomendaciones personalizadas.
-
-### **1.4 Alternativas Tecnológicas para Medir el Consumo**
-
-* **Descripción:**  
-   Evaluar tecnologías no tradicionales para medir el agua.
-
-* **Ejemplo:**  
-   Sensores acústicos, de presión, vibración o ultrasonidos.
-
-### **1.5 Optimización del Consumo en Dispositivos Domésticos**
-
-* **Descripción:**  
-   Automatizar el uso eficiente del agua en electrodomésticos y riego.
-
-* **Ejemplo:**  
-   Sistema domótico que regule duchas o lavadoras según patrones de consumo y disponibilidad.
-
-### **1.6 Gamificación y Concienciación**
-
-* **Descripción:**  
-   Incentivar el ahorro mediante dinámicas lúdicas.
-
-* **Ejemplo:**  
-   App móvil con recompensas por reducir el consumo respecto a períodos anteriores.
-
-### **1.7 Ciberseguridad y Privacidad de los Datos**
-
-* **Descripción:**  
-   Garantizar protección de datos personales y de consumo.
-
-* **Ejemplo:**  
-   Uso de encriptación, autenticación robusta o blockchain para trazabilidad.
-
-### **1.8 Integración con Sistemas de Smart City**
-
-* **Descripción:**  
-   Conectar MICA con plataformas urbanas para mejorar la gestión hídrica.
-
-* **Ejemplo:**  
-   APIs que permitan interoperar con sistemas municipales de agua.
-
-### **1.9 Mantenimiento Predictivo**
-
-* **Descripción:**  
-   Predecir fallos en la infraestructura mediante análisis histórico.
-
-* **Ejemplo:**  
-   Algoritmos que anticipen problemas en tuberías y optimicen mantenimientos.
-
-### **1.10 Visualización Avanzada y Realidad Aumentada**
-
-* **Descripción:**  
-   Hacer tangible el consumo hídrico con visualización interactiva.
-
-* **Ejemplo:**  
-   App con RA que muestre el flujo en tiempo real dentro del hogar.
-
----
-
-## **2\. 🧰 Infraestructura Básica Requerida**
-
-### **2.1 Componentes de Hardware**
-
-* **Sensor de Caudal:**  
-   Calibrado para medir consumo en la entrada de la vivienda.
-
-* **ESP32:**  
-   Unidad de procesamiento y transmisión.
-
-* **Wi-Fi:**  
-   Acceso a red para transmisión en tiempo real.
-
-### **2.2 Software y Plataformas de Comunicación**
-
-* **Broker MQTT:**  
-   Servidor de mensajes preconfigurado.
-
-* **Plataforma de Análisis:**  
-   Web y móvil con gráficos, alertas y visualizaciones.
-
-* **APIs de Integración:**  
-   Endpoints para facilitar nuevos módulos o servicios.
-
-### **2.3 Datos y Configuraciones Previas**
-
-* **Base de Datos Histórica:**  
-   Datos reales o simulados para pruebas y entrenamiento de modelos.
-
-* **Parámetros Iniciales:**  
-   Umbrales de alerta, consumo medio por región, etc.
-
-### **2.4 Infraestructura en la Nube**
-
-* **Servidor Cloud:**  
-   Capacidad para alojar web y base de datos, acceso remoto.
-
-* **Documentación Técnica:**  
-   Guías para conexión al MQTT, APIs y uso de la plataforma.
-
-### **2.5 Consideraciones Adicionales**
-
-#### **Componentes Preinstalados:**
-
-* Broker MQTT configurado.
-
-* Plataforma web y móvil operativa.
-
-* Base de datos con datos históricos.
-
-#### **Instalables por Participantes:**
-
-* Configuración de sensor en ESP32.
-
-* Instalación de nuevas librerías o módulos.
-
-#### **Soporte Técnico:**
-
-* Manuales y ejemplos de código para conectar sensores y plataformas.
-
----
-
-## **3\. 🧠 Democratización del Conocimiento: El Rol de los LLM en la Innovación Ciudadana**
-
-### **Introducción**
-
-Los modelos de lenguaje (LLM) están democratizando el acceso al conocimiento, permitiendo que cualquier persona, sin formación técnica avanzada, pueda crear soluciones de alto impacto. Este reto propone reflexionar sobre su papel como *mentores virtuales* en procesos de innovación ciudadana.
-
-### **Oportunidades y Retos**
-
-* **Empoderamiento del Ciudadano:**  
-   Asistentes virtuales personalizados para orientación técnica y conceptual.
-
-* **Reducción de Barreras de Conocimiento:**  
-   Accesibilidad a conceptos complejos explicados de forma sencilla.
-
-* **Innovación Colaborativa:**  
-   Co-creación y aprendizaje acelerado mediante herramientas abiertas y colaborativas.
-
-### **Ejemplos Abiertos para la Reflexión**
-
-1. **Mentoría Virtual Comunitaria:**  
-    Asesoramiento en tiempo real para proyectos ciudadanos.
-
-2. **Plataformas de Conocimiento Abierto:**  
-    Acceso universal a documentación, casos de éxito y mejores prácticas.
-
-3. **Interfaz Conversacional Técnica:**  
-    Resolución de problemas con lenguaje natural, sin conocimientos previos técnicos.
-
-### **Idea de Tool**
-
-[Ver idea de herramienta en Google Docs](https://docs.google.com/document/d/11j1nd5OV9GcWaq1-EoUupNpRmOKXRnvUyEbu-20Ubbs/edit)
-
-### **Conclusión**
-
-Los LLM pueden ser motores de innovación accesible, facilitando que personas con recursos limitados participen activamente en la creación de soluciones a retos sociales, como el consumo de agua.
-
----
-
-## **✅ Conclusión General**
-
-Este documento guía la participación en la **Hackatón MICA**, estableciendo tanto el marco tecnológico mínimo como los retos y áreas de innovación donde los equipos pueden aportar valor. La combinación de una infraestructura sólida y espacio creativo busca catalizar soluciones reales, colaborativas y escalables para la sostenibilidad hídrica.
-
