--- MICA/Hackaton UPF Mataro/documentos soporte/📚 Análisis No Intrusivo aplicado al Consumo Doméstico de Agua.md
@@ -1,118 +1,0 @@
-# **📚 Análisis No Intrusivo aplicado al Consumo Doméstico de Agua**
-
-## **1\. 📌 ¿Qué es el Análisis No Intrusivo?**
-
-El **Análisis No Intrusivo** (*Non-Intrusive Load Monitoring – NILM*) es una técnica analítica que permite identificar y clasificar distintos consumos individuales a partir de una sola medición agregada en un punto central, sin necesidad de medir directamente cada dispositivo o actividad por separado.
-
-* Originalmente desarrollado para identificar el consumo eléctrico individual de aparatos domésticos desde un único contador inteligente.
-
-* Usa técnicas avanzadas de procesamiento de señales y aprendizaje automático (Machine Learning e Inteligencia Artificial).
-
-## **2\. 📌 Aplicaciones iniciales del Análisis No Intrusivo (NILM)**
-
-La técnica NILM fue introducida inicialmente en el contexto energético:
-
-* **Electricidad**: Reconocer patrones característicos de aparatos eléctricos (lavadoras, frigoríficos, hornos, iluminación) mediante señales eléctricas medidas en el contador general.
-
-* El objetivo: optimizar el consumo energético, promover la eficiencia y mejorar la gestión de recursos.
-
-Ejemplo clásico:
-
-* Un contador inteligente mide la señal eléctrica total del hogar.
-
-* Algoritmos identifican eventos específicos (encendido/apagado electrodomésticos).
-
-* Clasificación automática usando algoritmos como SVM, Redes Neuronales, Redes Convolucionales (CNN), etc.
-
-## **3\. 📌 Enfoque del Análisis No Intrusivo para el Consumo de Agua**
-
-El enfoque NILM aplicado al consumo de agua sigue principios similares al eléctrico, adaptados al contexto hidráulico:
-
-* **Medición única del consumo total**: Lectura mediante un único contador inteligente que registra consumo instantáneo en intervalos breves (segundos o minutos).
-
-* **Identificación de patrones específicos**: Reconocer consumos individuales (ducha, cisterna, lavadora, lavavajillas, grifos) por patrones característicos de flujo y volumen.
-
-* Uso de datos en alta frecuencia (mediciones cada pocos segundos o minutos) para capturar detalles específicos del uso doméstico del agua.
-
-Ejemplos prácticos:
-
-* Ducha: Consumo sostenido con caudal relativamente estable durante algunos minutos.
-
-* Cisterna: Descarga breve de volumen alto seguida por un corte abrupto.
-
-* Lavadora: Ciclos periódicos con caudales variables durante cierto tiempo.
-
-## **4\. 📌 Estudios Científicos Relevantes en el Ámbito del NILM aplicado al Agua**
-
-Algunos estudios clave:
-
-* **"Non-intrusive Water Usage Classification Considering Limited Training Data"** (IEEE, 2020\)
-
-  * Propuesta: Generar datos sintéticos de eventos de consumo de agua cuando los datos reales son limitados.
-
-  * Algoritmo: Clasificación mediante series temporales sintéticas, enfocándose en eventos específicos del hogar (ducha, lavadora).
-
-  * [Enlace al estudio (IEEE)](https://ieeexplore.ieee.org/document/9054845)
-
-* **"Tracking Human Behavioural Consistency by Analysing Periodicity of Household Water Consumption"** (arXiv, 2019\)
-
-  * Análisis de patrones periódicos en series temporales medidas en hogares reales.
-
-  * Demuestra la capacidad de identificar hábitos regulares de consumo que facilitan clasificaciones automáticas.
-
-  * [Enlace al estudio (arXiv)](https://arxiv.org/abs/1905.05025)
-
-Estos estudios destacan la efectividad del NILM para caracterizar el consumo doméstico, incluso con limitaciones en la disponibilidad de datos reales.
-
-## **5\. 📌 Desafíos Técnicos y Científicos del NILM aplicado al Consumo de Agua**
-
-Aunque prometedora, la aplicación del NILM en el contexto del agua presenta desafíos específicos:
-
-### **🔸 Desafíos técnicos:**
-
-* **Resolución de datos**: Requiere datos con alta frecuencia (mediciones cada pocos segundos), lo que implica dispositivos más sofisticados y almacenamiento masivo.
-
-* **Superposición de eventos**: Dificultad para distinguir eventos simultáneos (ducha y lavadora funcionando al mismo tiempo).
-
-* **Ruido en señales hidráulicas**: El ruido en las mediciones (caídas de presión, variaciones de caudal inesperadas) complica la identificación precisa.
-
-### **🔸 Desafíos metodológicos:**
-
-* **Limitación de datos etiquetados**: Dificultad para obtener datasets grandes y bien etiquetados debido a privacidad y costos.
-
-* **Variabilidad en patrones domésticos**: Cada hogar tiene patrones únicos (número de habitantes, horarios, hábitos), complicando la generalización del modelo.
-
-### **🔸 Desafíos éticos y legales:**
-
-* **Privacidad de datos**: Información detallada del consumo podría revelar rutinas personales sensibles.
-
-* **Marco regulatorio**: Necesidad de cumplir normativas de privacidad (RGPD) en el uso y manejo de datos.
-
-## **6\. 📌 Estrategias para abordar estos desafíos**
-
-Para superar estas barreras, se pueden considerar varias estrategias científicas:
-
-* **Generación de datos sintéticos**:
-
-  * Crear conjuntos artificiales que simulen eventos de consumo doméstico basados en patrones conocidos (ayuda en ausencia de datos reales).
-
-* **Modelos híbridos (supervisados y no supervisados)**:
-
-  * Combinación de técnicas de aprendizaje supervisado (cuando hay datos etiquetados) y no supervisado para clasificar eventos en tiempo real con pocos datos etiquetados.
-
-* **Uso de modelos Deep Learning avanzados (CNN, LSTM)**:
-
-  * Modelos capaces de captar patrones complejos en series temporales hidráulicas, especialmente útiles para eventos superpuestos o ambiguos.
-
-* **Acuerdos colaborativos y éticos**:
-
-  * Establecer colaboraciones institucionales con suministradoras o entidades públicas para obtener datos anonimizados y cumplir con regulaciones de privacidad.
-
----
-
-## **🔖 Conclusión científica**
-
-El Análisis No Intrusivo aplicado al consumo doméstico de agua presenta un potencial significativo para mejorar la eficiencia hídrica en hogares mediante inteligencia artificial, aunque conlleva desafíos importantes tanto técnicos como éticos. La generación de datos sintéticos, el uso de modelos híbridos y colaboraciones institucionales son fundamentales para avanzar en esta prometedora línea de investigación.
-
-Este enfoque abre oportunidades significativas para el desarrollo de soluciones tecnológicas innovadoras orientadas al consumo responsable y sostenible del agua en entornos residenciales.
-
