--- MICA/Hackaton UPF Mataro/ejemplos codigo sensores/AS5600 magnetic encoder/platformio.ini
@@ -1,6 +1,0 @@
-[env:esp8266]
-platform = espressif8266
-board = nodemcuv2     ; o d1_mini si usas un Wemos D1 mini
-framework = arduino
-monitor_speed = 115200
-upload_speed = 115200
