--- MICA/Hackaton UPF Mataro/ejemplos codigo sensores/hall_sensor/platformio.ini
@@ -1,14 +1,0 @@
-[env:esp8266]
-platform = espressif8266
-board = nodemcuv2
-framework = arduino
-monitor_speed = 115200
-lib_deps = 
-  mprograms/QMC5883LCompass @ ^1.2.3
-
-
-[env:esp32-c6-devkitm-1]
-platform  = https://github.com/pioarduino/platform-espressif32/releases/download/51.03.04/platform-espressif32.zip
-board = esp32-c6-devkitm-1
-framework = arduino
-monitor_speed = 115200
