--- como usar github copilot.md
@@ -1,55 +1,0 @@
-# 🧭 Guía para Usar GitHub Copilot
-
-## Introducción
-GitHub Copilot es una herramienta de inteligencia artificial que ayuda a los desarrolladores a escribir código más rápido y con mayor eficiencia. Sin embargo, su uso no se limita solo a la generación de código.
-
-## Cómo Usar GitHub Copilot
-
-### Activación de GitHub Copilot
-
-#### Instalación del Plugin:
-Asegúrate de que tienes el plugin de GitHub Copilot instalado en tu editor de código. Actualmente, GitHub Copilot está disponible para Visual Studio Code, Visual Studio, Neovim y JetBrains.
-
-#### Inicio de Sesión:
-Inicia sesión en tu cuenta de GitHub desde el editor de código para activar GitHub Copilot.
-
-#### Configuración:
-Configura GitHub Copilot según tus preferencias. Puedes ajustar la cantidad de sugerencias que deseas recibir y la forma en que se presentan.
-
-### Uso de GitHub Copilot para Diversas Tareas
-
-#### Generación de Código:
-- Comienza a escribir el código como lo harías normalmente. GitHub Copilot te proporcionará sugerencias de completado automático en tiempo real.
-- Usa la tecla Tab para aceptar una sugerencia de GitHub Copilot. También puedes usar Ctrl + Space para ver más sugerencias.
-
-#### Análisis de Documentos:
-Utiliza GitHub Copilot para analizar documentos existentes. Puedes pedirle que resuma el contenido, explique conceptos o identifique puntos clave.
-
-#### Modificación de Documentos:
-Pide a GitHub Copilot que te ayude a modificar documentos. Puedes solicitarle que reescriba secciones para mejorar la claridad, corrija errores gramaticales o añada información adicional.
-
-#### Completación de Documentos:
-Si tienes documentos incompletos, GitHub Copilot puede ayudarte a completarlos. Proporciona el contexto necesario y deja que Copilot sugiera el contenido faltante.
-
-#### Desarrollo de Ideas:
-Usa GitHub Copilot para desarrollar nuevas ideas. Puedes describir un concepto y pedirle que sugiera implementaciones, ejemplos o mejoras.
-
-### Recomendaciones para Manejar el Contexto de las Preguntas
-
-#### Proporciona Contexto Detallado:
-Cuanto más contexto proporciones en tu pregunta, más precisas serán las sugerencias de GitHub Copilot. Incluye detalles sobre el objetivo de la tarea, el problema que estás tratando de resolver y otros aspectos relevantes.
-
-#### Usa Comentarios:
-Utiliza comentarios en tu código para guiar a GitHub Copilot. Los comentarios detallados pueden ayudar a Copilot a entender mejor lo que intentas lograr.
-
-#### Estructura Tus Preguntas:
-Estructura tus preguntas de manera clara y concisa. Evita preguntas ambiguas y proporciona toda la información relevante para obtener respuestas útiles.
-
-#### Revisa las Sugerencias:
-Siempre revisa las sugerencias proporcionadas por GitHub Copilot. Asegúrate de que el contenido cumple con tus requisitos y estándares de calidad.
-
-#### Ajusta las Configuraciones:
-Ajusta las configuraciones de GitHub Copilot según tus necesidades. Puedes cambiar la cantidad de sugerencias, el idioma de preferencia y más.
-
-## Conclusión
-GitHub Copilot es una herramienta poderosa que puede mejorar significativamente tu productividad en diversas tareas, no solo en la generación de código. Siguiendo estas recomendaciones, puedes aprovechar al máximo sus capacidades.
